+++ AWS-IoT-Arduino-Yun-Library/aws_iot_config_SDK.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2015 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2010-2016 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -19,6 +19,5 @@
 #define MAX_BUF_SIZE 256										// maximum number of bytes to publish/receive
 #define MAX_SUB 15 												// maximum number of subscribe
 #define CMD_TIME_OUT 100										// maximum time to wait for feedback from AR9331, 100 = 10 sec
-#define MAX_SHADOW_TOPIC_LEN 64                                 // maximum length for shadow topic, the metadata length for shadow topic is 32, make sure your thing name length plus that does not exceed this limit
 
 #endif