+++ AWS-IoT-Arduino-Yun-Library/examples/ThermostatSimulatorDevice/aws_iot_config.h
@@ -21,10 +21,10 @@
 #define AWS_IOT_MQTT_HOST "<CUSTOMER ENDPOINT HERE>" 	// your endpoint
 #define AWS_IOT_MQTT_PORT 8883									// your port
 #define AWS_IOT_CLIENT_ID	"My_ClientID"						// your client ID
-#define AWS_IOT_MY_THING_NAME "My_Board"						// your thing name
+#define AWS_IOT_MY_THING_NAME "room"						// your thing name
 #define AWS_IOT_ROOT_CA_FILENAME "aws-iot-rootCA.crt"           // your root-CA filename
-#define AWS_IOT_CERTIFICATE_FILENAME "cert.pem"                 // your certificate filename
-#define AWS_IOT_PRIVATE_KEY_FILENAME "privkey.pem"              // your private key filename
+#define AWS_IOT_CERTIFICATE_FILENAME "certificate.pem.crt"                 // your certificate filename
+#define AWS_IOT_PRIVATE_KEY_FILENAME "private.pem.key"              // your private key filename
 //===============================================================
 // SDK config, DO NOT modify it
 #define AWS_IOT_PATH_PREFIX "../certs/"