+++ AWS-IoT-Arduino-Yun-Library/examples/ThingShadowEcho/aws_iot_config.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2015 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2010-2016 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -18,7 +18,7 @@
 
 // Copy and paste your configuration into this file
 //===============================================================
-#define AWS_IOT_MQTT_HOST "data.iot.us-east-1.amazonaws.com" 	// your endpoint
+#define AWS_IOT_MQTT_HOST "<CUSTOMER ENDPOINT HERE>" 	// your endpoint
 #define AWS_IOT_MQTT_PORT 8883									// your port
 #define AWS_IOT_CLIENT_ID	"My_ClientID"						// your client ID
 #define AWS_IOT_MY_THING_NAME "My_Board"						// your thing name
@@ -27,7 +27,7 @@
 #define AWS_IOT_PRIVATE_KEY_FILENAME "privkey.pem"              // your private key filename
 //===============================================================
 // SDK config, DO NOT modify it
-#define AWS_IOT_PATH_PREFIX "./certs/"
+#define AWS_IOT_PATH_PREFIX "../certs/"
 #define AWS_IOT_ROOT_CA_PATH AWS_IOT_PATH_PREFIX AWS_IOT_ROOT_CA_FILENAME			// use this in config call
 #define AWS_IOT_CERTIFICATE_PATH AWS_IOT_PATH_PREFIX AWS_IOT_CERTIFICATE_FILENAME	// use this in config call
 #define AWS_IOT_PRIVATE_KEY_PATH AWS_IOT_PATH_PREFIX AWS_IOT_PRIVATE_KEY_FILENAME	// use this in config call