+++ AWS-IoT-Python-Runtime/runtime/run.py
@@ -1,0 +1,4 @@
+from runtimeHub import *
+
+AWSIoTMQTTArduinoPyHub = runtimeHub("AWSIoTMQTTArduinoHub", "../log/")
+AWSIoTMQTTArduinoPyHub.run()